Remove incorrect BurgerTest, add BurgerSetBunsTest, add BurgerIngredientTest, remove unused import statements
