+++ .idea/compiler.xml
@@ -6,6 +6,7 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
+        <module name="QA-java-diplom-1" />
         <module name="praktikum" />
       </profile>
     </annotationProcessing>
