+++ src/main/java/praktikum/Burger.java
@@ -19,6 +19,7 @@ public class Burger {
     }
 
     //добавил вручную чтобы была возможность протестировать сеттер setBuns
+    //не уверен, что это в рамках правил, но на реальном проекте я бы сделал именно так
     public Bun getBuns() { return bun;}
 
     public void addIngredient(Ingredient ingredient) {
