+++ src/test/java/praktikum/BurgerSetBunsTest.java
@@ -1,0 +1,21 @@
+package praktikum;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.Mock;
+import org.mockito.junit.MockitoJUnitRunner;
+import static org.junit.Assert.assertEquals;
+
+@RunWith(MockitoJUnitRunner.class)
+public class BurgerSetBunsTest {
+
+    @Mock
+    private Bun bun;
+
+    @Test
+    public void checkSetBuns() {
+        Burger burger = new Burger();
+        burger.setBuns(bun);
+        assertEquals(burger.getBuns(),bun);
+    }
+}
